Move client-related context components and hooks into src/client.tsx
